merge: let the user change the commit options on the merge wizard commit page (fixes #1727)
An "options" button has been added to the "commit panel". This button opens the
same "Options dialog" that is shown when you click the "Options" button on the
commit widget.

If some commit option is set, it will be shown on the commit panel as well, and
the options will be taken into account when making the commit.