filedata: perform binary, file size and line length checks when displaying patches
e8ada1a4d891 introduced a check for max line length to avoid hunging the
workbench when displaying binary files, which tend to have big "lines".

This test (and all other related tests, such as max file size and the binarity
check) were skipped when displaying pure patches. This happens when showing
unapplied patches and shelves. This result in hunging the workbench or the
shelve tool when displaying binary files. This patch fixes that problem.

Note that currently there is no simple way to tell TortoiseHg to display the
file anyway.

Also note that this code introduces a bit of code duplication, since the checks
are done on different places for patches and for regular revisions. I have not
seen an easy way to unify them since the revision check does a few things that
are unnecessary for patches.