repowidget: teach _openLink how to open relative repo paths
This will make it possible, for example, to open subrepos that are contained
within the current active repo (i.e. the repo of the current active tab) by
using their relative path rather than their absolute path.

This in turn will be used to be able to open repos by clicking on links on the
info bar.