repowidget: rename 'subrepo:' custom link type to 'repo:'
The existing 'subrepo:' custom link type is used throughout the code to make it
possible to open subrepos by providing their path. However, nothing in the code
is subrepo specific. That is, the existing 'subrepo:' link type can be used to
open a regular repository as well. In fact, aeed9101b546684c introduced such a
usage.

Thus, for consistency's sake, this revision renames the 'subrepo:' custom link
type to 'repo:' in order to better reflect its intent and usage.