infobar: convert subrepo references and subrepo revision id hashes into links
These links let you open the corresponding subrepo in a new workbench tab, and
even select the corresponding revision on the new subrepo workbench tab.

Note that this parses the mercurial output, which since 378f4b00e98d appends the
message '(in subrepo XXX)' after subrepo errors. Due to the way we parse the
mercurial output currently this only works for English messages.