infobar: get the subrepo path from the SubrepoAbort exception
Mercurial now adds an attribute containing the path of the failing subrepo to
SubrepoAbort exceptions.

We can use it to remove one of the regular expressions that we were using to
find subrepo error messages. In addition, this will work with subrepos whose
path contains spaces (which was not the case before). Finally, it should make it
easier to work with translated mercurial output (that is left for a future patch
though).

Note that with this change we will no longer be able to detect subrepo
references on non error messages. Thus we must remove the calls to
linkifyMessage from the non error infobar classes.