fileview, shelve: add 'Display the file anyway' link when files or patches are not shown
There are several cases in which the selected file (or patch) is not shown on
the fileview or on the shelve tool. For example, the file contents are not shown
when the file is binary, when it exceeds the maximum diff size or the maximum
line length, or a when it is a "large file".

This patch adds a 'Display the file anyway' link when the file is not shown for
one of the reasons above. When this link is clicked TortoiseHg will show the
file anyway. This change is temporary (i.e. the next time the file is selected
its contents will be hidden again and the link to force its display will be
shown again).

# Notes:

- The link will not be shown when the file is not shown due to an error.
- We have to be a bit careful with how we use the indicatorClicked signal, since
we can also have indicators used to show the excluded changes.
- There is a bit of code duplication between the fileview and the ChunksDialog.
The code is just different enough that we cannot just reuse the same code on
both classes.
- Technically the message is slightly inaccurate for patches (i.e. the righthand
side of the shelve dialog). I think that using a different message would
complicate the code too much for the little gain that we'd get.