phases: show a milder warning when changing phase from draft to secret
The existing warning does not make sense, since this phase change is generally
safe. Instead show a more focused warning that explains the rare cases in which
this transition is not 100% safe. Also, do not mention 'force' on the dialog,
despite the fact that mercurial requires --force for this transition, and change
the confirm button into 'Make secret'.