settings: focus on the default setting when opening settings from the workbench
The SettingsDialog already has a default setting that gets the focus if no
setting is selected by default. Do not select a different one (setting a
different default setting in two different locations) when the SettingsDialog is
open from the File menu of the Workbench.

This should make it easier to change the default setting in the future.

Note that this change also changes the default settings page when the settings
dialog is open from the workbench from the Workbench page to the TortoiseHg
page. This makes sense because the Workbench is the main window of the
TortoiseHg application.