settings: open a default settings page when the focus field does not exist
Up until now, if the selected 'focus' settings field is on a settings page that
is only enabled when an extension is active, or if there is a mistake on the
name of the focus field, no page is shown once the settings dialog is open.

This patch ensures that the Settings dialog will always open a known settings
page when this happens.