commit, revdetails: change splitter style to make it easier to see on Windows
The splitter between the commit message and the fileview on the commit and the
revision details panels is very hard to see on Windows. This patch changes its
style to 'Plastique' so that it has a visible handle.

Note that plenty of windows applications use this style, so it is not out of
place on a Windows application.