repowidget: manage connection of repository/configChanged signals in C++ layer
Thanks to @pyqtSlot, the slot function is exposed to C++ layer and its
connection will be managed in Qt/C++ world. Thus, it will be disconnected
immediately after the deletion of C++ object.

As we use w.deleteLater() and it occasionally generates half-alive object,
this change should be good workaround for "C/C++ object has been deleted"
error.

I have another plan for repositoryDestroyed, so it's kept untouched.

Tested with the following steps:

 1. while true; do touch .hg/store/00changelog.i; sleep 0.5; done
 2. repeat open/close of that repository