graph: only show "visible" revisions
This is necessary to be able to use the evolution extension productivelly.
Note that this is not the best way to fix this issue, but it is the simplest and
safest.

We should do a proper fix on default, using repo.filtered('visible') and
repo.changelog.revs on a for loop rather than using a while loop and increasing
the revision number one by one.

For mercurial versions earlier than 2.5 we revert to the old behavior.

Thanks to Pierre-Yves David for his help writing this patch and to Yuya for his
review.