revdetails: add revision details dialog
This dialog simply contains a RevDetailsWidget. It is meant to be able to show
the contents of a revision on an stand-alone dialog. This could be used, for
example, to show the revision contents of the revision corresponding to an
annotation dialog.

Currently you can only open this dialog by using the revdetails thg command.

Clicking on the parent/children links updates the dialog to show the
corresponding revision.

Clicking on a subrepo "Open..." link opens the subrepo on the workbench.