annotate: 'Go to Original revision' on Ctrl+click or double click on margin
This makes navigating to the originating revision very quick.

Since this uses setMarginSensitivity() and doing so disables the regular
"select line on margin click" behavior, this patch also tries to mimic that
behavior. The result is very similar but not exactly identical behavior when
Shift+clicking on the margin.

It would be nice to also change the mouse pointer to a link pointer when the
control key is held and the mouse is over the annotate margin, but this seems
difficult to do with the QScintilla API (the closest I found is
SCI_SETMARGINCURSORN, but this only lets you select between a regular or
reversed arrow cursor).