terminaltools: introduce terminaltools configuration settings
This is similar to the editortools. Key differences are that we support "args"
but not "argsln" or "argssearch" and that we can define windows, mac and unix
specific executables.

It is possible to use the string %(reponame)s on the executable or args
definitions, and it will be replaced by the repository name or path (depending
on where "Open Terminal" is called from). I decided to _not_ introduce a new
$REPONAME alias (akin to $FILE, $LINENUM, etc on editortools) because we did
not support that before. In fact we already suggested using %(reponame)s on the
tortoisehg.shell help on the settings dialog. We could introduce $REPONAME later
but we may need to handle the use of [] when setting tortosiehg.shell directly,
etc. For now I decided to just support what we already supported.

This patch also introduces a fair amount of shells. Notable missing shells are
MINGW and RXVT. Hopefully someone will add those later.