repowidget: show the actual number of outgoing changesets
A few people have come to me complaining that the "outgoing infobar" is a bit
confusing because it does not show the actual number of outgoing changesets,
which my differ from the total number of unpushed revisions when the default
push mode is not "All".

This patch tries to improve the situation, by indicating the number of revisions
that will be pushed in addition to the total number of outgoing revisions when
the "Push" button is clicked.

Note that we cannot simply show the revisions that will be pushed (and hide the
rest) because we want to let people override their default push action by right
clicking on one of the revisions that would not normally be pushed.

Also note that I discussed this with Pierre-Yves David on IRC and he gave me
some pointers to make this as fast as we could. Nevertheless this could have a
(hopefully) small impact on very large repos.

The code to create the "outgoing string" requires a long set of if/elif/else
clauses because it handles all possible cases individually in order to make the corresponding strings as translation friendly as possible.

In addition the Push button test will be changed to 'Nothing to push' and will
be disabled whenever there is nothing to push.