sync: remove URL splitting and -push and some password hiding
The intent is for the sync tool buttons to *always* use the URL shown in the
text entry.

The layout of the sync tool is now mostly static, items aren't shown and hidden
except for the options label which is usually not shown. Uses Mercurial's url
class effectively

TODO: add per-repo LRU history to the urllabel