messageentry: make reflowBlock markdown and "end of line period" aware
Up until now, using the Apply Format or the reflow keyboard shortcut on markdown
formatted text would completelly mess up the markdown text. This change makes
the reflow algorithm a bit "markdown" aware.

Since markdown is, at its core, a way to take text that uses very common
formatting conventions (such as using "-" to mark list elements, etc), most
commit messages are "markdown" formatted. Thus this makes using the reflow
algorithm on typical commit messages much better.

Also, the reflow algorithm will now respect "end of line periods", which means
that if a line ends with a perdiod, the next line will not be reflown into that
line. This makes it possible to type several separate statements on different
lines without the need to add extra, empty lines between them.

This change also has the _intended_ side effect of not reflowing the single line
first paragraphs (if it starts on the first line). If the first paragraph has
more than one line it will be reflown normally. In other words, summary lines
will not be reflown unless they already span more than one line. I believe this
is a best practice (i.e. summary lines should be shorter than 80 chars if
possible, but if that is not possible, it is better to have a longer summary
line than have it span more than one line). Since I think that this is a good
practice I believe that we should make it easy to follow. If a user does not
want this he can always manually split the summary line in two and then reflow
it.