repofilter: add "Show/Hide hidden changesets" toggle button
This includes a new SVG icon. I based the new icon on the existing hg-rebase
icon. It is a bit simple and could probably be improved but for now it can do.

It may make sense to rename repomodel.setBranch() into setShownRevisions() since
it now does more than just setting the branch.

It may also make sense to remember the "allparent" setting, which now is always
assumed to be False by default when the repository widget is created.