strip: allow stripping hidden revisions
Note that we have discussed that --hidden should only be passed to mercurial
commands when hidden revisions are shown. However in the case of strip we can
always pass it --hidden safely, since strip will always strip all the
descendants of a revision. Thus in this case --hidden will just let us choose a
hidden revision as the base revision to strip.