customtools: add a "Custom Tools" submenu to the commit file context dialogs
You can refer to the selected files on the custom tool definition by using the
"{SELECTEDFILES}" variable.

You can select the statuses of the files for which the custom tool will be shown
by setting the "status" setting of the custom tool definition. For now there is
no GUI for this.