revdetails: add filter editbox
This lets you filter which files are shown on the filelist of the revision
details. The match is a simple match (similar to the used by status.py).

When a new revision is selected the filter is reapplied, which means that the
new revision details file list may be empty (if no files match the filter in
the newly selected revision).

Due to space concerns I did not add a QLabel for old versions of QT as it is
done on the status widget.

It may make sense to hide the "Show All" button when it is not needed to
increase the available space.