customtools: add HookConfigDialog
This new config dialog is not used yet. It will be used by the new 'Hooks' panel
that will be introduced on the next revision.

This dialog is based on the recently added 'CustomConfigDialog' class.

The dialog will let the user configure the three hook settings, namely the hook
type, its name and the command that must be run when the hook is triggered.

The dialog has a 'Hook type' combobox which lists the hooks that are mentionned
on the hooks section of 'hg help config', and in the same order. The only hooks
that are not included are the 'pre-command' and 'post-command' hooks.