repomodel: use a gray color when displaying hidden revisions
The color is not as light as the unapplied patch color (because hidden revisions
are part of the repo, while unapplied patches are not). Also, the color is
applied to the descriptions as well to make them easier to spot on the DAG (this
is not necessary for unapplied patches because they are all at the top of the
repowidget).