customtools: add HooksFrame class
This new class defines a new settings frame that will be used to configure
repository hooks. Note that this is not enabled yet. It will be enabled on the
next revision, which will actually add the panel to the settings dialog.

The panel shows a list of hook configurations as a table, and buttons that allow
to add new hooks and edit or delete existing hooks. The table has three columns,
'Type', 'Name' and 'Command'.

If the user double cliks on one of the 'Type' items, a HookConvigDialog will be
shown, letting the user select the hook type, but also configure the hook name
and command. If the user double clicks on an item on one of the other colums the
user will be able to edit it inline.