thgrepo: update old uimtime only in check method
Previously uimtime was remembered as propertycache and cleared by
thginvalidate() or invalidateui(). So, it couldn't detect ui change in the
following case:

 1. "Refresh" -> thginvalidate() -> del _uimtime
 2. touch .hg/hgrc
 3. _checkuimtime() -> update _uimtime
    (mtime == _uimtime but it should mtime > _uimtime)

With this change, _uimtime is managed in the same manner as _repomtime, except
for the initial value. It's initialized by the current time for simplicity.

I'll remove deprecated thgrepository._uimtime later by follow-up patch.