repowidget: do show the MQ widget when a patch is applied or unapplied
The MQ widget used to be shown briefly (causing a small flicker) whenever a
patch was applied or unapplied. In addition, the MQ widget would sometimes be
selected, which was inconsistent with what happens when you select the top
applied patch.

Note that mqDemand's qgotoRevision() only works if the mq widget has already
been created. Thus we must call its get() method first, which creates the
widget if necessary and returns it. Only then we call qgotoRevision().