messageentry: fix Apply Format when paragraph has a single line (closes #2597)
We were not formatting any single line paragraphs, when only the first (subject)
paragraph should not be formatted.

If the first (subject) paragraph has a single line it is not formatted to ensure
that commit subjects have a single line by default. The user can still force
the formatting of the first line by manually making the first paragraph span
more than one line, in which case we apply the format to that first paragraph
as well.