repowidget: work around unwanted rejectBundle during pull to here (fixes #2596)
Because "pull to here" doesn't touch the confirmation infobar, the subsequent
pull operation clears the infobar, which involves rejectBundle(). As a result,
the incoming path changes before starting pull.

    rejectBundle -> reload -> restore original sync URL

This works around the problem by removing the confirmation explicitly. We'll
need to refactor handling of push/pull state later. This change is tricky.