customtools: add "Custom Tools" context menu to revdetails and manifest widgets
You can refer to the selected files on the custom tool definition by using the
"{SELECTEDFILES}" variable. Note that the file tree on the manifest widget
(currently) does not support multiple file selection but the revision details
widget does.

If the file(s) no longer exists on the working directory, a warning dialog will
be shown.

Also note that we do not (currently) support limiting the commands to different
mercurial statuses, as the file list on the commit widget does.