sync: add 'Update subrepo paths' checkbox to save dialog
When enabled, the selected path alias will also be created (or updated, if it
already exists) on all mercurial subrepos (i.e. not for non hg subrepos). This
is done by using the top repository path as the "base path", to which the local
relative subrepo path is appended.

If the mercurial subrepo has a trivial relative path defined on the .hgsub file
the change is performed without confirmation. If the subrepo path has an
absolute or non trivial relative path the use will be asked for confirmation.

For example, imagine a repository with two subrepos, located in "sub1" and
"myfolder/sub2" respectively. If the user creates a new "myserver" sync path on
the top repository, with URL "http://myserver.com/myproject", and enables this
new option, the following sync aliases will be created on those subrepos:

- Subrepo sub1:
    myserver = http://myserver.com/myproject/sub1
- Subrepo myfolder/sub2:
    myserver = http://myserver.com/myproject/myfolder/sub2

If the "myserver" sync path is then changed at the top repository to:
"C:\repos\mylocalclone" then the aliases will be updated to:

- Subrepo sub1:
    myserver = C:\repos\mylocalclone\sub1
- Subpreo myfolder/sub2:
    myserver = C:\repos\mylocalclone\myfolder\sub2

This option is applied recursively (i.e. of a subrepo has subrepos, those will
be updated as well).