settings: add 'never' option to 'Monitor Repo Changes' setting (refs #3227)
One of the work arounds suggested for issue #3227 was to set the 'Monitor Repo
Changes' to local only. However this may not work in some cases in which the
UNC path has been mapped to a windows drive. In those cases it may be handy to
completely disable the file system monitoring.