csinfo: add "obsolete" field
This field can show the obsolete "state" of a changeset. By "obsolete state" we
refer to any obsolete/changeset evolution related "flag", such as whether the
revision is obsolete, extinct or has some obsolete "trouble" (such as it being
divergent, bumped or unstable).

The list of "states" is shown as a list of strings, each with a light red
background.

This is not used yet, but it will be enabled in a following changeset.