revpanel: add ellipsis to summaries if the whole description spans several lines
This makes it easier to tell if a summary does not show the whole description.
This has two main benefits:

- It lets you tell if the summary on parent / child / precursor / successor
fields shows the complete description of the corresponding revision or not.

- It makes it more practical to collapse the description text area, giving more
space to the file/diff contents area. The reason this helps in this scenario is
that you can collapse the description area, and only uncollapse it briefly
whenever you see that the revision summary has an ellipsis at the end.