hglib: add ellipsis to longsummaries when it would have exceeded the limit
Up until now the longsummary function only added an ellipsis if the longsummary
mode was disabled (i.e. if no longsummary limit was set). This revision changes
that so that if the longsummary limit is set, and the longsummary would have
exceeded the limit, an ellipsis is added to indicate that there is text in the
description that is not included on the longsummary.

Note that in order to implement this change the longsummary() function has been
partially rewritten, with the intent of removing some duplicated code.

Another small side effect of the rewrite is that any indentation and empty lines
will be removed prior to generating the long summary, which makes the
longsummary include more useful text.